refactor : compareTo 생성
