+++ src/main/java/racingcar/model/Winners.java
@@ -3,6 +3,8 @@ package racingcar.model;
 import java.util.ArrayList;
 import java.util.List;
 
+import racingcar.model.dto.WinnerDTO;
+
 public final class Winners {
 	private List<Car> winners;
 
@@ -15,26 +17,26 @@ public final class Winners {
 	}
 
 	public void findWinners(final Cars cars) {
-		int maxDistance = 0;
+		Distance maxDistance = new Distance(0);
 		for (int carNum = 0; carNum < cars.size(); carNum++) {
 			WinnerDTO winnerDTO = new WinnerDTO(cars, maxDistance, carNum);
 			addCarTheSameSpeedAsFastest(winnerDTO);
-			maxDistance = addFasterCar(winnerDTO);
+			maxDistance = new Distance(addFasterCar(winnerDTO));
 		}
 	}
 
 	private void addCarTheSameSpeedAsFastest(WinnerDTO winnerDTO) {
-		if (winnerDTO.cars().car(winnerDTO.carNum()).distance().distance() == winnerDTO.maxDistance()) {
+		if (winnerDTO.cars().car(winnerDTO.carNum()).distance().equals(winnerDTO.maxDistance())) {
 			winners.add(winnerDTO.cars().car(winnerDTO.carNum()));
 		}
 	}
 
 	private int addFasterCar(WinnerDTO winnerDTO) {
-		if (winnerDTO.cars().car(winnerDTO.carNum()).distance().distance() > winnerDTO.maxDistance()) {
+		if (winnerDTO.cars().car(winnerDTO.carNum()).distance().compareTo(winnerDTO.maxDistance())) {
 			winners = new ArrayList<>();
 			winners.add(winnerDTO.cars().car(winnerDTO.carNum()));
-			winnerDTO.createMaxDistance(winnerDTO.cars().car(winnerDTO.carNum()).distance().distance());
+			winnerDTO.createMaxDistance(winnerDTO.cars().car(winnerDTO.carNum()).distance());
 		}
-		return winnerDTO.maxDistance();
+		return winnerDTO.maxDistance().distance();
 	}
 }
