+++ src/main/java/racingcar/model/dto/WinnerDTO.java
@@ -1,4 +1,7 @@
-package racingcar.model;
+package racingcar.model.dto;
+
+import racingcar.model.Cars;
+import racingcar.model.Distance;
 
 public class WinnerDTO {
 	private final Cars cars;
@@ -7,13 +10,13 @@ public class WinnerDTO {
 		return cars;
 	}
 
-	private int maxDistance;
+	private Distance maxDistance;
 
-	public int maxDistance() {
+	public Distance maxDistance() {
 		return maxDistance;
 	}
 
-	public void createMaxDistance(int maxDistance){
+	public void createMaxDistance(Distance maxDistance) {
 		this.maxDistance = maxDistance;
 	}
 
@@ -23,7 +26,7 @@ public class WinnerDTO {
 		return carNum;
 	}
 
-	public WinnerDTO(Cars cars, int maxDistance, int carNum) {
+	public WinnerDTO(Cars cars, Distance maxDistance, int carNum) {
 		this.cars = cars;
 		this.maxDistance = maxDistance;
 		this.carNum = carNum;
